requirements.txt
flask>=3.0.0
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-flask>=1.3.0
